duplicate folders fix
