+++ packages/manager/src/features/CloudPulse/Alerts/CreateAlert/NotificationChannels/AddNotificationChannel.tsx
@@ -1,5 +1,6 @@
 import { yupResolver } from '@hookform/resolvers/yup';
 import { Autocomplete, Box, Typography } from '@linode/ui';
+import { Grid } from '@mui/material';
 import React from 'react';
 import { Controller, FormProvider, useForm, useWatch } from 'react-hook-form';
 
@@ -8,6 +9,7 @@ import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { channelTypeOptions } from '../../constants';
 import { getAlertBoxStyles } from '../../Utils/utils';
 import { notificationChannelSchema } from '../schemas';
+import { RenderChannelDetails } from './RenderChannelDetails';
 
 import type { NotificationChannelForm } from '../types';
 import type { ChannelType, NotificationChannel } from '@linode/api-v4';
@@ -72,6 +74,15 @@ export const AddNotificationChannel = (props: AddNotificationChannelProps) => {
           (template) => template.channel_type === channelTypeWatcher
         )
       : null;
+
+  const templateOptions = React.useMemo(() => {
+    return selectedChannelTypeTemplate
+      ? selectedChannelTypeTemplate.map((template) => ({
+          label: template.label,
+          value: template.label,
+        }))
+      : [];
+  }, [selectedChannelTypeTemplate]);
 
   const selectedTemplate = selectedChannelTypeTemplate?.find(
     (template) => template.label === channelLabelWatcher
@@ -140,14 +151,18 @@ export const AddNotificationChannel = (props: AddNotificationChannelProps) => {
             <Controller
               render={({ field, fieldState }) => (
                 <Autocomplete
-                  onChange={(_, selected: { label: string }, reason) => {
+                  onChange={(
+                    _,
+                    selected: { label: string; value: string },
+                    reason
+                  ) => {
                     field.onChange(
-                      reason === 'selectOption' ? selected.label : null
+                      reason === 'selectOption' ? selected.value : null
                     );
                   }}
                   value={
-                    selectedChannelTypeTemplate?.find(
-                      (option) => option.label === field.value
+                    templateOptions.find(
+                      (option) => option.value === field.value
                     ) ?? null
                   }
                   data-testid="channel-label"
@@ -156,7 +171,7 @@ export const AddNotificationChannel = (props: AddNotificationChannelProps) => {
                   key={channelTypeWatcher}
                   label="Channel"
                   onBlur={field.onBlur}
-                  options={selectedChannelTypeTemplate ?? []}
+                  options={templateOptions}
                   placeholder="Select a Channel"
                 />
               )}
@@ -164,6 +179,25 @@ export const AddNotificationChannel = (props: AddNotificationChannelProps) => {
               name="label"
             />
           </Box>
+          {selectedTemplate && selectedTemplate.channel_type === 'email' && (
+            <Box paddingTop={2}>
+              <Grid container>
+                <Grid item md={1} sm={1} xs={2}>
+                  <Typography variant="h3">To:</Typography>
+                </Grid>
+                <Grid
+                  item
+                  md="auto"
+                  overflow="auto"
+                  paddingRight={1}
+                  sm="auto"
+                  xs="auto"
+                >
+                  <RenderChannelDetails template={selectedTemplate} />
+                </Grid>
+              </Grid>
+            </Box>
+          )}
         </Box>
         <ActionsPanel
           primaryButtonProps={{
